Fix : correct output format

+++ src/main/java/com/example/Shippable.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public interface Shippable {
 
     BigDecimal calculateShippingCost();
-    BigDecimal weight();
+    double weight();
 
 }
